Tickdata (Price) are decoded as float numbers (instead of decimal) to support bid/ask-values of  commodity futures.

+++ lib/ib/messages/incoming/ticks.rb
@@ -66,7 +66,7 @@ module IB
       TickPrice = def_message [1, 6], AbstractTick,
                               [:ticker_id, :int],
                               [:tick_type, :int],
-                              [:price, :decimal],
+                              [:price, :float],
                               [:size, :int],
                               [:can_auto_execute, :int]
 			class TickPrice
